Work on sauce

+++ controllers/sauces.js
@@ -3,7 +3,6 @@ const Sauce = require('../models/sauces');
 
 // récupère fonction fs pour pouvoir lire / modifier / mettre à jour les fichiers
 const fs = require('fs');
-const sauces = require('../models/sauces');
 
 // Création d'une sauce 
 exports.createSauce = (req, res, next) => {
@@ -29,53 +28,53 @@ exports.createSauce = (req, res, next) => {
 // Récupérer une seule sauce 
 exports.getOneSauce = (req, res, next) => {
   // On récupère l'id des paramètres de route
-    Sauce.findOne({
-      _id:req.params.id
-    })
+    Sauce.findOne({_id:req.params.id})
         .then(sauce => res.status(201).json({sauce}))
-        .catch(error => res.status(404).json({ error }));
+        //.catch(error => res.status(404).json({ error }));
+        .catch(error => res.status(404).json({ message: error }));
 
 };
+
+
 
 // Obtenir toutes les sauces - renvoie le tableau de toutes les sauces présentes dans la base de données
 exports.getAllSauces = (req, res, next) => {
   Sauce.find()
-    .then((sauces) => {
-      res.status(200).json(sauces);
-    })
-    .catch((error) => {
-      res.status(400).json({
-        error: error,
-      });
+    .then((sauces) => {res.status(200).json(sauces);})
+    .catch((error) => {res.status(400).json({error});
     });
 };
  
 
 
 // Modifier une sauce 
-exports.modifySauce = (req, res, next) => {
-    // si req.file existe (il y a déjà une image enregistrée mais on souhaite la modifiée), il faut le prendre en compte. On récupère avec JSON.parse les informations et on va générer l'image URL (car il s'agit d'une nouvelle image)
-    const sauceObject = req.file ? {
-      ...JSON.parse(req.body.sauce),
-      imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
-    } : { ...req.body };
-    
-    Sauce.updateOne({ _id: req.params.id }, { ...sauceObject, _id: req.params.id })
-      .then(() => res.status(200).json({ message: 'sauce modifiée !' }))
-      .catch(error => res.status(400).json({error}));
-};
+// si req.file existe (il y a déjà une image enregistrée mais on souhaite la modifiée), il faut le prendre en compte. On récupère avec JSON.parse les informations et on va générer l'image URL (car il s'agit d'une nouvelle image)
+exports.modifySauce = (req, res, _) => {
+  const sauceObject = req.file ? {
+    ...JSON.parse(req.body.sauce),
+    imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
+  } : { ...req.body };
+  
+  Sauce.updateOne({ _id: req.params.id }, { ...sauceObject, _id: req.params.id })
+    .then(() => res.status(200).json({ message: 'sauce modifiée !' }))
+    .catch(error => res.status(400).json({message : 'error modification'}));
+    };
 
 // Supprimer une sauce 
 exports.deleteSauce = (req, res, next) => {
     Sauce.findOne({ _id: req.params.id })
-      .then(sauce => {
+      .then((sauce) => {
+        if (sauce.userId !== req.user) {  // vérification que l'auteur de la sauce est bien celui qui souhaite la supprimer 
+          res.status(401).json({message: "action non autorisée"});  // si ce ne sont pas le cas (id différent) // non autorisé
+          return sauce;
+        }
         const filename = sauce.imageUrl.split('/images/')[1];
         fs.unlink(`images/${filename}`, () => {
           Sauce.deleteOne({ _id: req.params.id })
-            .then(() => res.status(200).json({ message: 'Sauce supprimée !'}))
-            .catch(error => res.status(400).json({ error }));
-        });
-      })
-      .catch(error => res.status(500).json({ error }));
+            .then(() => res.status(200).json({ message: 'sauce supprimée.'}))
+            .catch(error => res.status(400).json({ error }));  
+        })
+      });
   };
+
